Initial refactoring of the version control api ##rvc
