+++ libr/util/Makefile
@@ -5,7 +5,7 @@ CFLAGS+=-DR2_PLUGIN_INCORE -I$(TOP)/shlr
 PCLIBS=@LIBZIP@ @DL_LIBS@
 OBJS=mem.o unum.o str.o hex.o file.o range.o charset.o xdg.o rxml.o
 OBJS+=prof.o cache.o sys.o buf.o sys_w32.o ubase64.o base85.o base91.o
-OBJS+=list.o chmod.o graph.o event.o alloc.o donut.o print_code.o rvc.o
+OBJS+=list.o chmod.o graph.o event.o alloc.o donut.o print_code.o
 OBJS+=regex/regcomp.o regex/regerror.o regex/regexec.o uleb128.o rstr.o
 OBJS+=sandbox.o calc.o thread.o thread_sem.o thread_lock.o thread_cond.o thread_chan.o
 OBJS+=strpool.o bitmap.o time.o format.o pie.o print.o utype.o w32.o w32dw.o
@@ -15,6 +15,7 @@ OBJS+=udiff.o bdiff.o stack.o queue.o tree.o idpool.o assert.o bplist.o
 OBJS+=punycode.o pkcs7.o x509.o asn1.o asn1_str.o json_parser.o json_indent.o skiplist.o
 OBJS+=pj.o rbtree.o intervaltree.o qrcode.o vector.o skyline.o str_constpool.o str_trim.o
 OBJS+=ascii_table.o protobuf.o graph_drawable.o axml.o sstext.o new_rbtree.o token.o
+OBJS+=rvc.o rvc_git.o rvc_rvc.o
 
 ifeq (${HAVE_GPERF},1)
 OBJS+=d/ascii.o
